grocery-delivery-platform/backend/auth-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.95.2
uvicorn==0.22.0

# MongoDB
motor==3.1.1
pymongo==4.3.3

# Elasticsearch
elasticsearch==7.17.9

# Environment variables
python-dotenv==0.21.1

# Data validation
pydantic==1.10.7
email-validator==1.3.1

# JWT handling (for OAuth tokens)
python-jose[cryptography]==3.3.0

# Date handling
python-dateutil==2.8.2

# HTTP client (for OAuth)
httpx==0.24.1

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Password hashing
passlib[bcrypt]==1.7.4

# Form parsing
python-multipart==0.0.6 